Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0
python-docx==0.8.11
requests==2.31.0
